refactor(openapi): split ops paths into submodules [OPE-424]
